+++ src/bplus_tree_interior_page_header.c
@@ -57,6 +57,6 @@ void set_is_last_page_of_level_of_bplus_tree_interior_page(void* page, int is_la
 void print_bplus_tree_interior_page_header(const void* page, const bplus_tree_tuple_defs* bpttd_p)
 {
 	print_bplus_tree_page_header(page, bpttd_p->page_size);
-	printf("least_keys_page_id : %llu\n", (unsigned long long int)get_least_keys_page_id_of_bplus_tree_interior_page(page, bpttd_p));
+	printf("least_keys_page_id : %"PRIu64"\n", get_least_keys_page_id_of_bplus_tree_interior_page(page, bpttd_p));
 	printf("is_last_page_of_level : %d\n", is_last_page_of_level_of_bplus_tree_interior_page(page, bpttd_p));
 }
